Update WeatherFragment layout

+++ app/src/main/java/com/hagarsoft/weatherapp/data/Wind.java
@@ -1,6 +1,6 @@
 package com.hagarsoft.weatherapp.data;
 
 public class Wind {
-    public float speed;
-    public float deg;
+    public double speed;
+    public double deg;
 }
